utils: Deprecate dm_get_ekn_version()
This function looked for a file called EKN_VERSION with a version number
for the on-disk database layout. We haven't supported more than one
on-disk layout at a time since the precursor to the precursor to this
library, so we don't need to read this file anymore.

Instead, it's the caller's responsibility to make sure they are using
the correct version of the DModel library with the correct on-disk
database layout. This is made trivial with Flatpak and DModel's major
version number.

https://phabricator.endlessm.com/T22395
(cherry picked from commit 6b6a440391436a10e4558a0df41f3cf7b5f39eaa)
Signed-off-by: Emmanuele Bassi <ebassi@gnome.org>