Fix out of bounds array access in ekn-vfs.c
Previously, ekn_vfs_get_file_for_uri was testing whether elements in a
tokens array were null, without testing the size of the array. This
change adds one additional call to g_strsplit. The first is to separate
the URI scheme from the path, and the second is to separate tokens in
the path. Tokens are then assigned to variables for readability.

https://phabricator.endlessm.com/T23125