Correctly detect supported URIs in ekn_vfs_parse_name
With this change, ekn_vfs_parse_name uses the same behaviour as
ekn_vfs_get_file_for_uri to decide whether a URI scheme is supported. If
it is not, the implementation of parse_name from GLocalVfs is used
instead. This behaviour is consistent with the existing behaviour, since
ekn_vfs_get_file_for_uri also falls back to the implementation of
get_file_for_uri from GLocalVfs when the scheme is not supported.

This is necessary because parse_name and get_file_for_uri behave
differently in GLocalVfs for certain paths.

https://phabricator.endlessm.com/T23125